web/compfest-bin/src/app/src/requirements.txt
flask
flask_wtf
flask_sqlalchemy
flask_migrate
weasyprint==51
pycryptodome
babel